Crate problem111

Problem 111 (Primes with runs)

Solution detail

This is just a matter of using a careful implementation. For each digit d, and for each k = 10, 9, 8, ..., 0, generate all numbers containing exactly k occurrences of the digit d. Among all such numbers, find the sum of all the primes - if that sum is non-zero, then add it to a running total and break; otherwise, move onto the next k.

Functions

answer
Solve the problem, returning the answer as a String

main

repeated_digit_prime_sum
The sum of all of the prime n digit numbers in which digit d appears exactly k times.

solve
Find the sum of all primes with a given number of digits, each having a maximum number of a single repeated digit.

to_integer
Turn the given digits into a number